requirements.txt
streamlit
groq
PyPDF2
phidata
logging
duckduckgo-search
